Fix memory leak in my cJSON usage
Thanks to valgrind on openbsd, found and fixed a few leaks here.

Main issue was in how the "extra" json metadata was being used.
Turns out you need to delete the data in the function it's
created.

This now adds the assumption that the JSON is properly converted
to serialized byte arrays before the logEventWithExtra() function
returns so it's safe to wipe the backing JSON memory.