No way to configure telemetry settings
This provides an initial stub-out of functionality to expand upon
later, but for now it adds a toggle switch in the setup tui app to
easily enable/disable telemetry. The config setting ties into a global
flag used by a macro added to the x_events.c functions to check if
they should really fire or not.

Shortcuts were taken, specifically cJSON objects tend to still be
created and destroyed needlessly as the only check for if the service
is enabled or not is done during init, close, and the writing routines.