No way to configure telemetry modes
Since the next step will be to add in UDP support, we need a way to
configure it and toggle between modes. This updates the setup dialog
to allow toggling modes (filesystem or udp) and configuring basic UDP
details (host, port).

Still needs some CLI arg logic, but that'll be later.