Lack of framework for adding in other loggers
This is a pretty large change that makes some things better, some
things maybe worse? Since the idea is to support multiple pluggable
frameworks for logging telemetry events (filesyste & udp as top
priority), this abstracts out a Logger "type" that is a collection of
management functions (init, close) and a writing function (write) that
can be generically called by the core logic that builds and serializes
JSON.

It also refactors and changes how the file system logging behaves,
which simplifies initialization.

Instead of individual files per level, it now creates a single file at
application start and only writes to it for the lifetime of the
app. This is to bring it in line with potential UDP broadcasting
behavior. The upside is it dramatically simplifies the horrid
initialization code that was woven into multiple places in the Doom
source and is now centralized in d_main.c where other
subsystems (video, network, etc.) are initialized.