Refactor some event values, fix win32 test
While working on a server side receiver, decided to refactor some of the
actor types to be more enum-like. This simplifies some coding at least
on the Java receiver side of things.

Also simplified the Move logging since there's no difference in logging
a player or enemy at the moment.

Lastly, fixed the test code to work properly on Win32/msys2.