We don't support a click-bait worthy message system :-)
Clearly we need more Kafka in our lives...this stubs out support for
using the `librdkafka` library (if available) to use as a target for
publishing telemetry events. Still needs meat on the bones, but gets
the config hooks ready.

Also updates the docs, variables, and tests to support
yet-another-mode.