Test fixes and (sadly) more ifdefs
Getting a bit out of hand, but properly got the ifdefs in place for
making the Kafka stuff optional for setup and actual doom binary.

Also moved the test building/execution logic into the Automake files,
which was very odd at first, but seems to be working. Big step forward
and should fix multi-platform test support since it's now managed by
Automake.

Was able to validate both the doom, setup, and test binaries build
properly with and without librdkafka configured.