Add more references to Player for certain events
In order to properly track Player state we should be including
references to the Player as actor for more events. For example,
for level start or end. The reason for this is to help track key
player state like position, health, and armor. (Health and armor
were not tracked before this commit.)

Even in the case of an enemy killing the player, the target is
now set to the player instance so we can track where the player
was at time of death.