Targets lacked Positions, caused processing issues
There's no reason why we can't include a current position for a target
and it makes things simpler for downstream processing of the data
since an actor and target become symetrical in terms of
schema.

Additionally this does have a beneficial side effect: it's possibly
someone Targets another actor that crosses a subsector boundary it's
very noticeable.