Improve the autoconf libtls and libbsd detection
Change the logic around for detecting libtls and make sure we try to
detect libbsd as well.

For libbsd, rely on first a function check (to detect if we're on a
native bsd system) and if it fails try pkg-config.

For libtls, first try pkg-config. If it fails, try finding the library
outside pkg-config, implying we're on a libressl-native system like
OpenBSD. (If so, manually append to LDFLAGS). Otherwise fail message
burps out.

I'll need to update the dumb websockets library, too.