win midi: Always prepare/unprepare the buffer before modifying
From Win API docs:

> After the header has been prepared, do not modify the buffer. After the driver
> is done using the buffer, call the midiOutUnprepareHeader function.

However, the docs also say this:

> The application can re-use the same buffer, or allocate multiple buffers and
> call midiOutPrepareHeader for each buffer. If you re-use the same buffer, it
> is not necessary to prepare the buffer each time. You can call
> midiOutPrepareHeader once at the beginning and then call
> midiOutUnprepareHeader once at the end.

Since ZMusic (GZDoom) and Portmidi (PrBoom+/DSDA-Doom) still call the
midiOutUnprepareHeader, we decided to call it as well. Also, we still have a
bug report with the occasional playback stuck. We can't reproduce it reliably,
but it seems to have been fixed after this change.